+++ packages/lightning-mock/src/types.ts
@@ -5,8 +5,9 @@ import type {
   Credential,
 } from '@openfn/lexicon/lightning';
 import type { ServerState } from './server';
+import { PhoenixEvent } from './socket-server';
 
-export type LightningEvents = 'log' | 'run-complete';
+export type LightningEvents = 'log' | 'run-complete' | string; // not complete!
 
 export type DevServer = Koa & {
   state: ServerState;
@@ -20,13 +21,15 @@ export type DevServer = Koa & {
   getQueueLength(): number;
   getResult(runId: string): any;
   getState(): ServerState;
+  messageSocketClients(message: PhoenixEvent): void;
   on(event: LightningEvents, fn: (evt: any) => void): void;
   once(event: LightningEvents, fn: (evt: any) => void): void;
   onSocketEvent(
     event: LightningEvents,
     runId: string,
-    fn: (evt: any) => void
-  ): void;
+    fn: (evt: any) => void,
+    once?: boolean
+  ): () => void;
   registerRun(run: LightningPlan): void;
   removeAllListeners(): void;
   reset(): void;